fix the lint failures blocking the gate
Two in the new code: the zero width space in the sift test was a literal,
which staticcheck reads as an accident rather than the point of the test, so
it is spelled as an escape now. The document count in the pass was converted
to the type it already had.

Five were already on main and fail the same gate: labelled, characterised and
colour in the card use text and its test.

+++ sach/pass.go
@@ -303,7 +303,7 @@ func (p *Pass) part(ctx context.Context, line *Line, from kho.Stored, done bool)
 		return c, t, err
 	}
 
-	c.Documents = int64(t.phoi.Documents)
+	c.Documents = t.phoi.Documents
 	c.Kept = int64(out.Documents())
 	c.Text = out.Text()
